+++ tests/buildtest/runner.rs
@@ -70,10 +70,7 @@ macro_rules! test_prepare_error {
         #[test]
         fn $name() {
             runner::run($krate, |run| {
-                let res = run.run(
-                    rustwide::cmd::SandboxBuilder::new().enable_networking(false),
-                    |_| Ok(()),
-                );
+                let res = run.run(crate::utils::sandbox_builder(), |_| Ok(()));
                 match res.err().and_then(|err| err.downcast().ok()) {
                     Some(rustwide::PrepareError::$expected) => {
                         // Everything is OK!
@@ -103,10 +100,7 @@ macro_rules! test_prepare_error_stderr {
         #[test]
         fn $name() {
             runner::run($krate, |run| {
-                let res = run.run(
-                    rustwide::cmd::SandboxBuilder::new().enable_networking(false),
-                    |_| Ok(()),
-                );
+                let res = run.run(crate::utils::sandbox_builder(), |_| Ok(()));
                 match res.err().and_then(|err| err.downcast().ok()) {
                     Some(rustwide::PrepareError::$expected(output)) => {
                         assert!(output.contains($expected_output), "output: {:?}", output);